+++ R/ecDistance.R
@@ -28,22 +28,34 @@ ecDistance <- function(x) {
 #-----------------------------------------------------------------------------------------------------------------------------------------------#
   
   if (class(x)[1] %in% c('SpatialPoints', 'SpatialPointsDataFrame')) {
-    od <- function(n) {
-     return(do.call(rbind, lapply(n, function(i) {
+    od <- function(y) {
+      n <- length(y)
+     return(do.call(cbind, lapply(1:n, function(i) {
        p1 <- x[i,]@coords
-       md <- sapply(n, function(p) {
+       md <- sapply(1:n, function(p) {
          p2 <- x[p,]@coords
          d <- sqrt((p1[1]-p2[1])^2 + (p1[2]-p2[2])^2)
          return(d)})})))}}
   
   if (class(x)[1] %in% c('SpatialPolygons', 'SpatialPolygonsDataFrame')) {
-    od <- function(n) {
-      return(do.call(rbind, lapply(n, function(i) {
-      p1 <- x[i,]
-      md <- sapply(n, function(p) {
-        p2 <- x[p,]
+    od <- function(y) {
+      n <- length(y)
+      return(do.call(cbind, lapply(1:n, function(i) {
+      p1 <- y[i,]
+      md <- sapply(1:n, function(p) {
+        p2 <- y[p,]
         d <- gDistance(p1, p2)
         return(d)})})))}}
+  
+  if (class(x)[1] %in% c('matrix', 'data.frame')) {
+    od <- function(y) {
+      n <- nrow(y)
+      return(do.call(cbind, lapply(1:n, function(i) {
+        p1 <- y[i,]
+        md <- sapply(1:n, function(j) {
+          p2 <- y[j,]
+          d <- sqrt((p1[1]-p2[1])^2 + (p1[2]-p2[2])^2)
+          return(d)})})))}}
   
 #-----------------------------------------------------------------------------------------------------------------------------------------------#
 # 2. build distance matrix
@@ -51,7 +63,7 @@ ecDistance <- function(x) {
   
   if (exists("od")) {
     
-    od <- od(1:length(x)) # estimate distance
+    od <- od(x) # estimate distance
     od[which(od==0)] <- NA # set redundant values to NA
     return(od)