+++ R/smoothTrace.R
@@ -16,19 +16,13 @@
 #' # read reference data
 #' data(fieldData)
 #' 
-#' # extract centroids for each polygon
-#' cp <- spCentroid(fieldData)
-#' 
-#' # build single polygon from centroid points (simple)
-#' p1 <- simpleTrace(cp)
-#' 
 #' # build polygon from raster (smooth)
-#' p2 <- smoothTrace(rasterize(p1, raster(extent(p1), res=30, crs=crs(p1))), 1)
+#' ref <- raster(extent(fieldData[1,]), res=30, crs=crs(fieldData))
+#' p <- smoothTrace(rasterize(fieldData[1,], ref), 1)
 #' 
 #' # compare objects
-#' plot(p1)
-#' points(cp, col="red")
-#' plot(p2, add=TRUE, border="blue")
+#' plot(fieldData[1,])
+#' plot(p, add=TRUE, border="blue")
 #' 
 #' }
 #' @export